move kanban to screens/projects, break out ticket dropdown fields stylesheet
