+++ src/screens/Project/TaskManagement/styles/ConfirmationDialogs.scss
