+++ src/screens/Project/TaskManagement/styles/NoTicketsCreated.scss
